feat: 이미지 파일 경로 디코딩 처리
utf-8로 기본 설정되어있지만, 이미지 리소스에 한글이 인코딩되어 있었음,
클라이언트에서 다시 한글로 디코딩하여 서버에 이미지 파일 삭제를 요청.

+++ routes/files.js
@@ -185,10 +185,14 @@ router.post('/images', imgUpload.array('img'), function (req, res, next) {
 
 router.delete('/images', async function (req, res, next) {
   try {
+    res.status(404).send(new Error('errrrrr'));
+    return;
     const { filePathList } = req.body;
-    const fileList = filePathList.split(',');
+    const fileList = JSON.parse(filePathList);
     for (const file of fileList) {
-      await fs.promises.unlink(file);
+      if (fs.existsSync(file)) {
+        await fs.promises.unlink(file);
+      }
     }
 
     res.status(200).send(true);
